+++ packages/search-adapter/src/empathy/__fixtures__/responses/single-facet.response.ts
@@ -1,6 +1,6 @@
 import { EmpathyFacet } from '../../models';
 
-export const deepFacet: EmpathyFacet = {
+export const hierarchicalRawFacet: EmpathyFacet = {
   'facet': 'hierarchical_category',
   'values': [
     {
@@ -13,18 +13,20 @@ export const deepFacet: EmpathyFacet = {
       'value': 'Juegos de manualidades',
       'count': 39,
       'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades"',
-      'selected': true,
+      'selected': false,
       'values': [
         {
           'value': 'Construye',
           'count': 39,
-          'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye"',
+          'filter':
+            '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye"',
           'selected': false,
           'values': [
             {
               'value': 'Construye Hija',
               'count': 10,
-              'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye/construye-hija"',
+              'filter':
+                '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye/construye-hija"',
               'selected': false
             }
           ]
@@ -34,7 +36,7 @@ export const deepFacet: EmpathyFacet = {
   ]
 };
 
-export const deepFacetWithNoSelectedProperty: EmpathyFacet = {
+export const hierarchicalRawFacetWithoutSelected: EmpathyFacet = {
   'facet': 'hierarchical_category',
   'values': [
     {
@@ -48,14 +50,22 @@ export const deepFacetWithNoSelectedProperty: EmpathyFacet = {
       'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades"',
       'values': [
         {
+          'value': 'Playmobil',
+          'count': 4,
+          'filter':
+            '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/playmobil"'
+        },
+        {
           'value': 'Construye',
-          'count': 39,
-          'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye"',
+          'count': 35,
+          'filter':
+            '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye"',
           'values': [
             {
               'value': 'Construye Hija',
               'count': 10,
-              'filter': '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye\\/construye-hija"'
+              'filter':
+                '{!tag=hierarchical_category}hierarchical_category:"juegos_de_manualidades\\/construye\\/construye-hija"'
             }
           ]
         }
@@ -64,7 +74,31 @@ export const deepFacetWithNoSelectedProperty: EmpathyFacet = {
   ]
 };
 
-export const priceFacet = {
+export const simpleRawFacet: EmpathyFacet = {
+  'facet': 'collection_facet',
+  'values': [
+    {
+      'count': 33,
+      'filter': '{!tag=collection_facet}collection_facet:"Hold"',
+      'selected': false,
+      'value': 'Hold'
+    },
+    {
+      'count': 20,
+      'filter': '{!tag=collection_facet}collection_facet:"Icon\\/Mesh"',
+      'selected': false,
+      'value': 'Icon Mesh'
+    },
+    {
+      'count': 14,
+      'filter': '{!tag=collection_facet}collection_facet:"Mesh\\/Color"',
+      'selected': false,
+      'value': 'Mesh Color'
+    }
+  ]
+};
+
+export const dynamicNumberRangeRawFacet: EmpathyFacet = {
   'facet': 'price_facet',
   'values': [
     {
@@ -106,6 +140,22 @@ export const priceFacet = {
       'value': '100:*',
       'count': 0,
       'filter': '{!tag=price_facet}priceSort:[100 TO *]'
+    }
+  ]
+};
+
+export const noDynamicNumberRangeRawFacet: EmpathyFacet = {
+  'facet': 'rating_facet',
+  'values': [
+    {
+      'value': '1:5',
+      'count': 11,
+      'filter': '{!tag=rating_facet}rating:[1 TO 5]'
+    },
+    {
+      'value': '5:10',
+      'count': 14,
+      'filter': '{!tag=rating_facet}rating:[5 TO 10]'
     }
   ]
 };