+++ packages/search-adapter/src/empathy/__mocks__/empathy-builder.mocks.ts
@@ -1,7 +1,9 @@
 import { inject, injectable, multiInject, optional } from 'inversify';
 import { DEPENDENCIES } from '../container/container.const';
 import { BeforeRequest, BeforeResponseTransform, RequestMapper, Requestor, ResponseTransformed } from '../empathy-adapter.types';
-import { EmpathyResultMapper, EmpathySimpleRequestMapper, ResponseMappers } from '../mappers';
+import { EmpathySimpleRequestMapper } from '../mappers/request/empathy-simple-request.mapper';
+import { ResponseMappers } from '../mappers/response.mappers';
+import { EmpathyResultMapper } from '../mappers/response/results/empathy-result.mapper';
 
 @injectable()
 export class CustomResultMapper extends EmpathyResultMapper {}