+++ packages/search-adapter/src/empathy/mappers/request/params/empathy-request-filters-solr-syntax.mapper.ts
@@ -1,4 +1,4 @@
-import { Filter, FilterModel, MultiSelect, RangeFilter, SimpleFilter } from '@empathy/search-types';
+import { Filter, MultiSelect, NumberRangeFilter, RangeValue } from '@empathy/search-types';
 import { inject, injectable } from 'inversify';
 import { EmpathyAdapterConfig, FacetConfig, FilterValueMapperParams } from '../../../config/empathy-adapter-config.types';
 import { DEPENDENCIES } from '../../../container/container.const';
@@ -15,80 +15,66 @@ export class EmpathyRequestFiltersSolrSyntaxMapper implements Mapper<Filter[], s
   constructor(@inject(DEPENDENCIES.config) private readonly config: EmpathyAdapterConfig) {}
 
   map(filters: Filter[], filterValues: string[]): string[] {
-    const facetName = filters[0].facet.modelName;
-    const mappedFilterValues = this.needsCompositionWithOrOnFrontend(facetName, filters)
-      ? this.composeFiltersWithOr(facetName, filters)
-      : this.composeFilters(facetName, filters);
+    const facetId = filters[0].facetId;
+    const mappedFilterValues = this.needsCompositionWithOrOnFrontend(facetId, filters)
+      ? this.composeFiltersWithOr(facetId, filters)
+      : this.composeFilters(facetId, filters);
     filterValues.push(...mappedFilterValues);
     return filterValues;
   }
 
-  private composeFilters(facetName: string, filters: Filter[]): string[] {
-    const mapFilterValue = this.getMappingFilterValueFn(filters[0].modelName);
-    return filters.map(filter =>
-      `${ this.getFilterPrefix(facetName, filter) }:${ mapFilterValue(filter) }`);
+  private composeFilters(facetId: string, filters: Filter[]): string[] {
+    return this.isNumberRangeFilters(filters)
+      ? filters.map(filter => `${ this.getFilterPrefix(facetId, filter) }:${ this.mapRangeValue(filter.value) }`)
+      : filters.map(filter => `${ this.getFilterPrefix(facetId, filter) }:${ filter.value }`);
   }
 
-  private composeFiltersWithOr(facetName: string, filters: Filter[]): string[] {
-    const prefix = this.getFilterPrefix(facetName, filters[0]);
-    const filterValues = filters.map(this.getMappingFilterValueFn(filters[0].modelName)).join(' OR ');
-    return [`${ prefix }:(${ filterValues })`];
+  private composeFiltersWithOr(facetId: string, filters: Filter[]): string[] {
+    const prefix = this.getFilterPrefix(facetId, filters[0]);
+    const filterValues = this.isNumberRangeFilters(filters)
+      ? filters.map(filter => this.mapRangeValue(filter.value))
+      : filters.map(filter => filter.value);
+    return [`${ prefix }:(${ filterValues.join(' OR ') })`];
   }
 
-  private getFacetConfig(facetName: string): FacetConfig {
+  private getFacetConfig(facetId: string): FacetConfig {
     const facetsConfig = this.config.mappings.facets;
-    return facetsConfig.named[facetName] || facetsConfig.default;
+    return facetsConfig.named[facetId] || facetsConfig.default;
   }
 
-  private getFacetNamePrefix(facetName: string, context: FilterValueMapperParams): string {
-    const { prefix } = this.getFacetConfig(facetName);
-    return prefix.facetName instanceof Function
-      ? prefix.facetName(context)
-      : prefix.facetName;
+  private getFacetIdPrefix(facetId: string, context: FilterValueMapperParams): string {
+    const { prefix } = this.getFacetConfig(facetId);
+    return prefix.facetId instanceof Function
+      ? prefix.facetId(context)
+      : prefix.facetId;
   }
 
-  private getFilterDeepness(filter: Filter, deepness: number = 0): number {
-    return filter.parent ? this.getFilterDeepness(filter.parent, deepness + 1) : deepness;
+  private getFilterPrefix(facetId: string, filter: Filter): string {
+    const { showUnselectedValues } = this.getFacetConfig(facetId);
+    const context: FilterValueMapperParams = { config: this.config, filter };
+    const noTagPrefix = showUnselectedValues ? this.getNoTagPrefix(facetId, context) : '';
+    const facetIdPrefix = this.getFacetIdPrefix(facetId, context);
+    return `${ noTagPrefix }${ facetIdPrefix }`;
   }
 
-  private getFilterPrefix(facetName: string, filter: Filter): string {
-    const { showUnselectedValues } = this.getFacetConfig(facetName);
-    const filterDeepness = this.getFilterDeepness(filter);
-    const context: FilterValueMapperParams = { config: this.config, filter, filterDeepness, facetName };
-    const noTagPrefix = showUnselectedValues ? this.getNoTagPrefix(facetName, context) : '';
-    const facetNamePrefix = this.getFacetNamePrefix(facetName, context);
-    return `${ noTagPrefix }${ facetNamePrefix }`;
+  private isNumberRangeFilters(filters: Filter[]): filters is NumberRangeFilter[] {
+    return filters[0].modelName === 'NumberRangeFilter';
   }
 
-  private getMappingFilterValueFn(model: string): (filter: Filter) => string {
-    switch (model) {
-      case FilterModel.currencyRange:
-      case FilterModel.range:
-        return this.mapRangeValue;
-      case FilterModel.simple:
-      default:
-        return this.mapSimpleValue;
-    }
+  private mapRangeValue({ min, max }: RangeValue): string {
+    return `[${ min || '*' } TO ${ max || '*' }]`;
   }
 
-  private getNoTagPrefix(facetName: string, context: FilterValueMapperParams): string {
-    const { prefix } = this.getFacetConfig(facetName);
-    const noTag = prefix.noTagFacetName instanceof Function
-      ? prefix.noTagFacetName(context)
-      : prefix.noTagFacetName;
+  private getNoTagPrefix(facetId: string, context: FilterValueMapperParams): string {
+    const { prefix } = this.getFacetConfig(facetId);
+    const noTag = prefix.noTagFacetId instanceof Function
+      ? prefix.noTagFacetId(context)
+      : prefix.noTagFacetId;
     return `{!tag=${ noTag }}`;
   }
 
-  private mapRangeValue({ value }: RangeFilter): string {
-    return `[${ value.min || '*' } TO ${ value.max || '*' }]`;
-  }
-
-  private mapSimpleValue({ value }: SimpleFilter): string {
-    return value.filter;
-  }
-
-  private needsCompositionWithOrOnFrontend(facetName: string, filters: Filter[]): boolean {
-    const { multiSelectable } = this.getFacetConfig(facetName);
+  private needsCompositionWithOrOnFrontend(facetId: string, filters: Filter[]): boolean {
+    const { multiSelectable } = this.getFacetConfig(facetId);
     return multiSelectable === MultiSelect.OnFrontend && filters.length > 1;
   }
 }