+++ packages/search-adapter/src/empathy/mappers/response/index.ts
@@ -1,5 +1,11 @@
+export * from './facets/empathy-facet.mapper';
+export * from './facets/empathy-hierarchical-facet.mapper';
+export * from './facets/empathy-number-range-facet.mapper';
+export * from './facets/empathy-simple-facet.mapper';
+
 export * from './filters/empathy-filter.mapper';
-export * from './filters/empathy-range-filter.mapper';
+export * from './filters/empathy-hierarchical-filter.mapper';
+export * from './filters/empathy-number-range-filter.mapper';
 export * from './filters/empathy-simple-filter.mapper';
 
 export * from './results/empathy-result.mapper';
@@ -9,7 +15,6 @@ export * from './suggestions/empathy-suggestion.mapper';
 export * from './suggestions/empathy-suggestion-facets.mapper';
 
 export * from './empathy-banner.mapper';
-export * from './empathy-facet.mapper';
 export * from './empathy-next-query.mapper';
 export * from './empathy-partial-result.mapper';
 export * from './empathy-promoted.mapper';